requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2
psycopg2-binary
google-api-python-client
google-auth-oauthlib